Closes #10. Add router for product and direct navigation.
Add a `backbone` router for product and direct navigation.

Had to revert back to `fragment` based routing as direct navigation was seriously messed up with `push state`
based routing. In the `onbeforeunload` event it is impossible to get access to the target url due to privacy
concerns (you have no idea where the user is going => they could be navigating to another site).

The ajax request for the page fires and of course being a single page application that resource is not found.
I might see later if there is a way to rewrite the url behind the scenes but maintain the correct browser
history.

Rant over.